Add tacacs mapped user support
This enables sudo for tacacs mapped users for libtacplus_map
That is, users logging under tacacs authentication with their
"real" name, but mapped to tacacs0...tacacs15 local users

Ideally, this should be done as a different sudo plugin, but it
would be an exact match for the default sudoers plugin, with
very small changes, and would therefore be a maintenance nightmare
to keep in sync.   Absent that, should be a configure change.  Maybe
at some point in the future.

Added libtaplus_map dependencies to debian/control.
Applied debian/patches files and then removed them.

Added some config files to .gitignore for ease of building directly
with git